pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "tableau2pbir"
version = "0.1.0"
description = "Deterministic Tableau (.twb/.twbx) to Power BI (PBIR) converter"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
dependencies = [
  "anthropic>=0.34,<1.0",
  "pydantic>=2.5,<3.0",
  "lxml>=5.0",
  "tableaudocumentapi>=0.11",
  "PyYAML>=6.0",
  "sqlglot>=23,<27",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-xdist>=3.5",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.scripts]
tableau2pbir = "tableau2pbir.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tableau2pbir"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
